Node/Nat status card MVP #2
