+++ src/Pages/Authenticated/Dashboard/NodeStatus/Bubble.scss
@@ -7,6 +7,7 @@
   border-radius: 50%;
   border: 5px solid;
   display: inline-block;
+
   &--failed {
     border-color: #CB0D0D;
   }
