+++ src/Pages/Authenticated/Dashboard/NodeStatus/NatStatus.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/base/variables';
+@import 'src/assets/styles/base/variables';
 
 .nat-status {
   display: flex;
