+++ src/Pages/Authenticated/Dashboard/NodeStatus/NodeStatus.scss
@@ -1,4 +1,4 @@
-@import '../../../../assets/styles/base/variables';
+@import 'src/assets/styles/base/variables';
 
 .status-card {
   padding: 22px 0 25px 28px;
@@ -8,13 +8,6 @@
   display: flex;
   align-items: center;
 
-  //&__status-text {
-  //  font-style: normal;
-  //  font-weight: $medium-weight;
-  //  font-size: 16px;
-  //  line-height: 24px;
-  //  color: $main-text-color;
-  //}
   &__status-text {
     font-style: normal;
     font-weight: $normal-weight;
